chore: move source files to the ./src/ folder
