--- components/ColorSchemeToggle/ColorSchemeToggle.tsx
@@ -1,15 +1,0 @@
-'use client';
-
-import { Button, Group, useMantineColorScheme } from '@mantine/core';
-
-export function ColorSchemeToggle() {
-  const { setColorScheme } = useMantineColorScheme();
-
-  return (
-    <Group justify="center" mt="xl">
-      <Button onClick={() => setColorScheme('light')}>Light</Button>
-      <Button onClick={() => setColorScheme('dark')}>Dark</Button>
-      <Button onClick={() => setColorScheme('auto')}>Auto</Button>
-    </Group>
-  );
-}
