+++ src/app/layout.tsx
@@ -1,7 +1,9 @@
-import '@mantine/core/styles.css';
 import React from 'react';
-import { MantineProvider, ColorSchemeScript } from '@mantine/core';
-import { theme } from '../theme';
+
+import { MantineProvider } from '@mantine/core';
+import '@mantine/core/styles.css';
+
+import { theme } from '@/theme';
 
 export const metadata = {
   title: 'ArrowFlicks',
@@ -12,7 +14,6 @@ export default function RootLayout({ children }: { children: any }) {
   return (
     <html lang="en">
       <head>
-        <ColorSchemeScript />
         <link rel="shortcut icon" href="/favicon.svg" />
         <meta
           name="viewport"
