+++ src/app/page.tsx
@@ -1,0 +1,9 @@
+import { Text } from '@mantine/core';
+
+export default function HomePage() {
+  return (
+    <>
+      <Text>Movies</Text>
+    </>
+  );
+}
